requirements.txt
flask
flask-socketio
paho-mqtt
gevent
gevent-websocket
